ai-core/scripts/train.py
#!/usr/bin/env python
# =============================================================================
# Model Training Script
# =============================================================================
"""
Train pronunciation scoring model.

Usage:
    python train.py --config configs/train_config.yaml
"""

import argparse
import logging
from pathlib import Path
from typing import Dict, Any

import yaml
import torch
import torch.nn as nn
from torch.utils.data import DataLoader
from tqdm import tqdm

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


def load_config(config_path: Path) -> Dict[str, Any]:
    """Load training configuration."""
    with open(config_path, "r") as f:
        return yaml.safe_load(f)


def setup_experiment(config: Dict[str, Any]) -> None:
    """Setup MLflow/W&B experiment tracking."""
    try:
        import mlflow
        mlflow.set_tracking_uri(config["logging"]["mlflow"]["tracking_uri"])
        mlflow.set_experiment(config["logging"]["mlflow"]["experiment_name"])
        mlflow.start_run(run_name=config["experiment"]["name"])
        mlflow.log_params(config)
        logger.info("MLflow experiment initialized")
    except ImportError:
        logger.warning("MLflow not available, skipping experiment tracking")


class PronunciationDataset(torch.utils.data.Dataset):
    """Dataset for pronunciation scoring."""
    
    def __init__(self, manifest_path: Path, config: Dict[str, Any]):
        self.manifest_path = manifest_path
        self.config = config
        self.samples = []
        
        # Load manifest
        if manifest_path.exists():
            import json
            with open(manifest_path / "manifest.json", "r") as f:
                self.samples = json.load(f)
        
        logger.info(f"Loaded {len(self.samples)} samples from {manifest_path}")
    
    def __len__(self) -> int:
        return len(self.samples)
    
    def __getitem__(self, idx: int) -> Dict[str, torch.Tensor]:
        sample = self.samples[idx]
        
        # Placeholder - actual implementation would load audio
        audio = torch.randn(16000)  # 1 second of fake audio
        phonemes = torch.zeros(50)  # Placeholder phoneme labels
        
        return {
            "audio": audio,
            "phonemes": phonemes,
            "text": sample.get("text", ""),
        }


class PronunciationScorer(nn.Module):
    """
    Pronunciation scoring model using Wav2Vec2 backbone.
    
    Architecture:
    - Wav2Vec2 encoder for audio features
    - Phoneme classifier head
    - GOP (Goodness of Pronunciation) regressor
    """
    
    def __init__(self, config: Dict[str, Any]):
        super().__init__()
        self.config = config
        
        # Placeholder model architecture
        self.encoder = nn.Sequential(
            nn.Linear(16000, 1024),
            nn.ReLU(),
            nn.Linear(1024, 256),
        )
        
        self.phoneme_head = nn.Linear(256, 44)  # CMU phoneme set
        self.gop_head = nn.Linear(256, 1)
    
    def forward(self, audio: torch.Tensor) -> Dict[str, torch.Tensor]:
        features = self.encoder(audio)
        
        phoneme_logits = self.phoneme_head(features)
        gop_scores = self.gop_head(features)
        
        return {
            "phoneme_logits": phoneme_logits,
            "gop_scores": gop_scores,
        }


def train_epoch(
    model: nn.Module,
    dataloader: DataLoader,
    optimizer: torch.optim.Optimizer,
    device: torch.device,
) -> Dict[str, float]:
    """Train for one epoch."""
    model.train()
    total_loss = 0.0
    
    for batch in tqdm(dataloader, desc="Training"):
        audio = batch["audio"].to(device)
        
        optimizer.zero_grad()
        outputs = model(audio)
        
        # Placeholder loss
        loss = outputs["gop_scores"].mean()
        
        loss.backward()
        optimizer.step()
        
        total_loss += loss.item()
    
    return {"train_loss": total_loss / len(dataloader)}


def validate(
    model: nn.Module,
    dataloader: DataLoader,
    device: torch.device,
) -> Dict[str, float]:
    """Validate model."""
    model.eval()
    total_loss = 0.0
    
    with torch.no_grad():
        for batch in tqdm(dataloader, desc="Validation"):
            audio = batch["audio"].to(device)
            outputs = model(audio)
            loss = outputs["gop_scores"].mean()
            total_loss += loss.item()
    
    return {"val_loss": total_loss / len(dataloader)}


def main():
    parser = argparse.ArgumentParser(description="Train pronunciation model")
    parser.add_argument(
        "--config",
        type=Path,
        default=Path("configs/train_config.yaml"),
        help="Path to training configuration",
    )
    
    args = parser.parse_args()
    
    # Load configuration
    config = load_config(args.config)
    logger.info(f"Loaded configuration: {config['experiment']['name']}")
    
    # Setup experiment tracking
    setup_experiment(config)
    
    # Setup device
    device = torch.device(
        "cuda" if torch.cuda.is_available() and config["training"]["device"] == "cuda"
        else "cpu"
    )
    logger.info(f"Using device: {device}")
    
    # Create datasets
    train_dataset = PronunciationDataset(
        Path(config["data"]["train_path"]),
        config,
    )
    val_dataset = PronunciationDataset(
        Path(config["data"]["val_path"]),
        config,
    )
    
    # Create dataloaders
    train_loader = DataLoader(
        train_dataset,
        batch_size=config["training"]["batch_size"],
        shuffle=True,
        num_workers=config["training"]["num_workers"],
    )
    val_loader = DataLoader(
        val_dataset,
        batch_size=config["training"]["batch_size"],
        shuffle=False,
        num_workers=config["training"]["num_workers"],
    )
    
    # Create model
    model = PronunciationScorer(config).to(device)
    logger.info(f"Model parameters: {sum(p.numel() for p in model.parameters()):,}")
    
    # Setup optimizer
    optimizer = torch.optim.AdamW(
        model.parameters(),
        lr=config["training"]["learning_rate"],
        weight_decay=config["training"]["weight_decay"],
    )
    
    # Training loop
    best_val_loss = float("inf")
    patience_counter = 0
    
    for epoch in range(config["training"]["epochs"]):
        logger.info(f"Epoch {epoch + 1}/{config['training']['epochs']}")
        
        train_metrics = train_epoch(model, train_loader, optimizer, device)
        val_metrics = validate(model, val_loader, device)
        
        logger.info(f"Train Loss: {train_metrics['train_loss']:.4f}")
        logger.info(f"Val Loss: {val_metrics['val_loss']:.4f}")
        
        # Save best model
        if val_metrics["val_loss"] < best_val_loss:
            best_val_loss = val_metrics["val_loss"]
            patience_counter = 0
            
            save_path = Path(config["checkpoint"]["save_dir"]) / "best.pt"
            save_path.parent.mkdir(parents=True, exist_ok=True)
            torch.save(model.state_dict(), save_path)
            logger.info(f"Saved best model to {save_path}")
        else:
            patience_counter += 1
            
        # Early stopping
        if patience_counter >= config["training"]["early_stopping"]["patience"]:
            logger.info("Early stopping triggered")
            break
    
    logger.info("Training complete!")


if __name__ == "__main__":
    main()
