ai-core/requirements.txt
# =============================================================================
# VietSpeak AI - AI Core Dependencies
# =============================================================================

# Core ML
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# Speech Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
webrtcvad==2.0.10
praat-parselmouth==0.4.3

# ASR Models
transformers==4.37.0
openai-whisper==20231117
faster-whisper==1.0.0

# Pronunciation Assessment
g2p-en==2.1.0
phonemizer==3.2.1

# Data Processing
pandas>=2.0.0
pyarrow>=14.0.0
datasets>=2.15.0

# Experiment Tracking
mlflow>=2.9.0
wandb>=0.16.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
hydra-core>=1.3.0
omegaconf>=2.3.0
